Remove intersphinx config and fix link to bugs (#25)
Intersphinx is no longer used in the docs building process.
Removing it from config will allow to build docs locally or for
the plugin itself without any issues.

A key for the bugs link in the extlinks dictionary should be unique
across all the plugins. It's important to specify it in platform docs
config. But in plugin it's useful to have it as well: to have an ability
to build docs for the plugin only without platform ones.