Ignore invalid paragraphs in package indices
If the long-description of a package contains an erroneous newline, the
rest of the paragraph is considered to describe a new package, but it is
incomplete. We try to ignore those and hope for the best.
Usually there is only the 'Homepage' field, so no functionality should be
affected.

(cherry picked from commit 75da379e46a0991043ed5cd349d5731fafbd389e)