Problem: plugin does not manage travis config with plugin-template
Solution: generate travis config and template_config.yml

This patch introduces the template_config.yml that helps manage the Travis config
for the plugin.

The config can be updated by running the following from the root of the plugin_template
repository:

./plugin-template --travis pulp_deb

Any features of the Travis config that are currently disabled can be enabled by updating
the template_config.yml file in this repository. After updating the file, the
./plugin-template --travis command needs to be run again.

[noissue]