tests: Add initial source package tests
[noissue]

related to:
https://github.com/pulp/pulp_deb/issues/409
(was https://pulp.plan.io/issues/8775)

The 'test_sync' test which already does sync tests on a remote repo
with and without installer (udeb) packages is extended to perform a
new round of testing which will include source packages.

This is a minimum set of tests for the new source package
support. Future commits will expand the testing of source package
support.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>